moved the panic function to io.rs, like the other boards
